+++ hw3/src/Shape.java
@@ -1,7 +1,8 @@
 public abstract class Shape {
-
+    //private int height; add later?
+    //private int width;
     abstract double area();
     abstract double perimeter();
-    abstract double getWidth();
-    abstract double getHeight();
+    abstract int getWidth();
+    abstract int getHeight();
 }
